# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: PlayBall API
  description: |
    **PlayBall** - 스포츠 매칭 플랫폼 REST API 명세서

    ## 인증 방식
    카카오 OAuth 로그인 후 서버에서 JWT를 발급합니다.
    - `POST /api/auth/kakao` → 카카오 인가코드로 로그인 → accessToken + refreshToken 발급
    - 인증 필요 API: `Authorization: Bearer {accessToken}` 헤더 포함
    - 비로그인 허용 API: `security: []` 표기

    ## 공통 응답
    ```json
    { "success": true, "message": "성공", "data": { ... } }
    ```

    ## 에러 응답
    ```json
    { "success": false, "message": "에러 메시지", "data": null }
    ```

    ## 종목 코드
    `SOCCER` | `BASKETBALL` | `RUNNING` | `BADMINTON`

    ## 실력 코드
    `BEGINNER` | `INTERMEDIATE` | `ADVANCED`
  version: 1.1.0
  contact:
    name: PlayBall Team
servers:
  - url: http://localhost:8080
    description: 로컬 개발
  - url: https://api.playball.kr
    description: 운영
...
security:
  - BearerAuth: []

# ============================================================
paths:
...
  /api/notifications:
    get:
      tags: [알림]
      summary: 내 알림 목록 조회 (Cursor 페이지네이션)
      description: |
        무한 스크롤 형태. 첫 호출은 cursor 생략, 응답의 `nextCursor`를 다음 호출에 전달.
        `nextCursor`가 null이면 더 이상 데이터가 없습니다.

        **알림 타입 (noticeType):**
        - `MATCH_FOUND`: 매칭 성사 / 참가 수락
        - `APPLICATION_REJECTED`: 참가 거절
        - `MATCH_REMINDER`: 경기 시작 1시간 전 리마인더
        - `MATCH_CANCELLED`: 경기 취소
        - `RATING_REQUEST`: 경기 종료 후 칭찬 요청
        - `SYSTEM_NOTICE`: 시스템 공지
      parameters:
        - name: cursor
          in: query
          description: 이전 응답의 nextCursor 값 (첫 호출 시 생략)
          schema:
            type: integer
            format: int64
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: onlyUnread
          in: query
          description: true이면 안 읽은 알림만 조회
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "알림 목록 조회 성공"
                data:
                  items:
                    - notificationId: 145
                      noticeType: "MATCH_FOUND"
                      title: "매칭 성사!"
                      content: "백석대 컴공 내기 매칭이 성사되었습니다."
                      isRead: false
                      targetType: "MATCH"
                      targetId: 12
                      createdAt: "2026-05-17T19:00:00"
                    - notificationId: 144
                      noticeType: "MATCH_REMINDER"
                      title: "경기 임박"
                      content: "오후 7시 '백석대 축구' 경기가 곧 시작됩니다."
                      isRead: true
                      targetType: "MATCH"
                      targetId: 1
                      createdAt: "2026-05-17T18:00:00"
                  nextCursor: 144
...
  /api/notifications/{notificationId}/read:
    patch:
      tags: [알림]
      summary: 알림 읽음 처리
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 읽음 처리 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "읽음 처리되었습니다"
                data: null
...
  /api/notifications/{notificationId}:
    delete:
      tags: [알림]
      summary: 알림 삭제
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 삭제 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "삭제되었습니다"
                data: null
  /api/notifications/settings:
    get:
      tags: [알림]
      summary: 알림 설정 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "알림 설정 조회 성공"
                data:
                  enabled: true
                  updatedAt: "2026-05-17T18:30:00"

    patch:
      tags: [알림]
      summary: 알림 ON/OFF 토글
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
            example:
              enabled: false
      responses:
        '200':
          description: 설정 변경 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "알림 설정이 변경되었습니다"
                data:
                  enabled: false
                  updatedAt: "2026-05-17T19:01:00"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "카카오 로그인 후 발급받은 accessToken"
...
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
...